refactor: move loose JS files to frontend/js/
