+++ frontend/js/loading-bar.js
