+++ frontend/js/simple-template.js
